+++ app/Http/Controllers/IngredientController.php
@@ -15,13 +15,14 @@ class IngredientController extends Controller
      */
     public function index()
     {
-        return view('ingredients.list',['ingredient' => Ingredient::orderBy('name')->get()]);
+        return view('ingredients.list', ['ingredient' => Ingredient::orderBy('name')->get()]);
     }
 
     public function show(Ingredient $ingredient)
     {
-        return view('ingredients.show',['ingredient'=> $ingredient]);   
+        return view('ingredients.show', ['ingredient' => $ingredient]);
     }
+
     /**
      * Show the form for creating a new resource.
      *
@@ -29,12 +30,13 @@ class IngredientController extends Controller
      */
     public function create()
     {
-        return view('ingredients.create');    }
+        return view('ingredients.create');
+    }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
@@ -43,21 +45,26 @@ class IngredientController extends Controller
             'name' => $request->input('name'),
             'stock' => $request->input('stock'),
         ];
+        $ingExist = Ingredient::where('name',$data['name'])->get();
+        if($ingExist){
+            return back()->with('error','L\'ingredient '.$data['name'].' existe deja dans la liste');
+        }
         $addIng = Ingredient::create($data);
-        return redirect()->route('ingredients.index')->with('success','Your ingredient '.$addIng->name.' was added to the liste!');    
+
+        return redirect()->route('ingredients.index')->with('success', 'Your ingredient ' . $addIng->name . ' was added to the liste!');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit(Ingredient $ingredient)
@@ -68,18 +75,18 @@ class IngredientController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $ingredient)
     {
         $modIng = Ingredient::find($ingredient);
-        
+
         $data = [
             'stock' => $request->input('stock'),
         ];
-        
+
         $modIng->update($data);
 
         return redirect()->route('ingredients.index');
@@ -88,21 +95,24 @@ class IngredientController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy(Ingredient $ingredient)
     {
-        if($ingredient->boisson->count() > 0){
+        if ($ingredient->name === 'Sucre') {
+            return back()->with('error', 'Can\'t delete this ingredient');
+        }
+        if ($ingredient->boisson->count() > 0) {
             $liste = '';
             $boissons = $ingredient->boisson;
-            foreach ($boissons as $boisson){
-                $liste .= "<li>" . $boisson->name ."; </li>";
+            foreach ($boissons as $boisson) {
+                $liste .= "<li>" . $boisson->name . "; </li>";
             }
             $info = "Can't delete this Ingredient. Associated with drinks: " . $liste;
             return back()->with('error', $info);
         }
         $ingredient->delete();
-        return redirect()->route('ingredients.index')->with('warning','The ingredient '. $ingredient->name .' has been deleted');    
+        return redirect()->route('ingredients.index')->with('warning', 'The ingredient ' . $ingredient->name . ' has been deleted');
     }
 }
